K-means Algorithm For clustering
Let's dive into the steps of implementation of segregating users on their interests;

So, each users interest will be stored in the data base.

We can  use K-means clustering to put people with similar methods into the same cluster.

Convert the dataset into numerical values if it is not already, so that calculations can be performed.

The User interest will be the cluster centroid.

All the remaining users interests arrays will be considered as datapoints.

The model would calculate the distance between the datapoint & all the centroids, and will be assigned to the cluster with the nearest centroid.

The user whose datapoints are matching with the centroid will  be put together in the same cluster ( Distance will be minimum).

How it Works

Initialize centroids by first shuffling the dataset and then randomly selecting K data points for the centroids without replacement.

Keep iterating until there is no change to the centroids. i.e assignment of data points to clusters isn’t changing.

Compute the sum of the squared distance between data points and all centroids.

Assign each data point to the closest cluster (centroid).